Fixed and refactored crypt test
