+++ tests/unit/Crypt/EncryptCest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Phalcon\Test\Unit\Crypt;
 
 use UnitTester;
+use Phalcon\Crypt;
 
 /**
  * Class EncryptCest
@@ -30,6 +31,35 @@ class EncryptCest
     public function cryptEncrypt(UnitTester $I)
     {
         $I->wantToTest('Crypt - encrypt()');
-        $I->skipTest('Need implementation');
+        $tests   = [
+            md5(uniqid())            => str_repeat('x', mt_rand(1, 255)),
+            time() . time()          => str_shuffle('abcdefeghijklmnopqrst'),
+            'le$ki12432543543543543' => "",
+        ];
+        $ciphers = [
+            'AES-128-ECB',
+            'AES-128-CBC',
+            'AES-128-CFB',
+            'AES-128-OFB',
+            'AES128',
+        ];
+
+        $crypt = new Crypt();
+        foreach ($ciphers as $cipher) {
+            $crypt->setCipher($cipher);
+
+            foreach ($tests as $key => $test) {
+                $crypt->setKey(substr($key, 0, 16));
+                $encryption = $crypt->encrypt($test);
+                $actual     = rtrim($crypt->decrypt($encryption), "\0");
+                $I->assertEquals($test, $actual);
+            }
+
+            foreach ($tests as $key => $test) {
+                $encryption = $crypt->encrypt($test, substr($key, 0, 16));
+                $actual     = rtrim($crypt->decrypt($encryption, substr($key, 0, 16)), "\0");
+                $I->assertEquals($test, $actual);
+            }
+        }
     }
 }
