+++ tests/unit/Crypt/SetCipherCest.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 namespace Phalcon\Test\Unit\Crypt;
 
 use UnitTester;
+use Phalcon\Crypt;
+use Phalcon\Crypt\Exception;
 
 /**
  * Class SetCipherCest
@@ -30,6 +32,31 @@ class SetCipherCest
     public function cryptSetCipher(UnitTester $I)
     {
         $I->wantToTest('Crypt - setCipher()');
-        $I->skipTest('Need implementation');
+
+        $cipher = "aes-256-cfb";
+        $crypt = new Crypt();
+        $crypt->setCipher($cipher);
+        $I->assertEquals($cipher, $crypt->getCipher());
+    }
+
+    /**
+     * Tests the Crypt::setCipher
+     *
+     * @author                   Phalcon Team <team@phalconphp.com>
+     * @since                    2018-05-06
+     *
+     * @expectedException        \Phalcon\Crypt\Exception
+     * @expectedExceptionMessage The cipher algorithm "xxx-yyy-zzz" is not
+     *                           supported on this system.
+     */
+    public function shouldThrowExceptionIfCipherIsUnknown(UnitTester $I)
+    {
+        $I->expectThrowable(
+            Exception::class,
+            function () {
+                $crypt = new Crypt();
+                $crypt->setCipher('xxx-yyy-zzz');
+            }
+        );
     }
 }
