+++ tests/unit/Crypt/SetPaddingCest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Phalcon\Test\Unit\Crypt;
 
 use UnitTester;
+use Phalcon\Crypt;
 
 /**
  * Class SetPaddingCest
@@ -25,11 +26,46 @@ class SetPaddingCest
      * @param UnitTester $I
      *
      * @author Phalcon Team <team@phalconphp.com>
-     * @since  2018-11-13
+     * @since  2014-10-17
      */
     public function cryptSetPadding(UnitTester $I)
     {
         $I->wantToTest('Crypt - setPadding()');
-        $I->skipTest('Need implementation');
+        $texts   = [''];
+        $key     = '0123456789ABCDEF0123456789ABCDEF';
+        $ciphers = [
+            'AES-256-ECB',
+            'AES-256-CBC',
+            'AES-256-CFB',
+        ];
+        $pads    = [
+            Crypt::PADDING_ANSI_X_923,
+            Crypt::PADDING_PKCS7,
+            Crypt::PADDING_ISO_10126,
+            Crypt::PADDING_ISO_IEC_7816_4,
+            Crypt::PADDING_ZERO,
+            Crypt::PADDING_SPACE,
+        ];
+
+        for ($i = 1; $i < 128; ++$i) {
+            $texts[] = str_repeat('A', $i);
+        }
+
+        $crypt = new Crypt();
+        $crypt->setKey(substr($key, 0, 32));
+
+        foreach ($pads as $padding) {
+            $crypt->setPadding($padding);
+
+            foreach ($ciphers as $cipher) {
+                $crypt->setCipher($cipher);
+
+                foreach ($texts as $expected) {
+                    $encrypted = $crypt->encrypt($expected);
+                    $actual    = $crypt->decrypt($encrypted);
+                    $I->assertEquals($expected, $actual);
+                }
+            }
+        }
     }
 }
